refactor: command를 view 레이어로 이동
