+++ src/main/java/blackjack/view/InputView.java
@@ -1,6 +1,5 @@
 package blackjack.view;
 
-import blackjack.controller.Command;
 import blackjack.model.betting.BettingMoney;
 import java.util.Arrays;
 import java.util.List;
